Declare Python 3.4 as incompatible (#807)
Python 3.4 testing has been dropped in #608, but it was not declared as
incompatible in setup.py because the code base was expected to still
work with Python 3.4 for a while. This compatibility finally broke with
the 21.1.0 release. At least one of the problematic commits is
e09b1d642339d7d94ba58d7855da40e65ed1a88f introducing an import of
the typing module not available in Python 3.4.